requirements.txt
Django>=5.0,<6.0
Pillow>=10.0
django-allauth>=0.57.0
psycopg2-binary>=2.9.0
gunicorn>=21.0
whitenoise>=6.0
